Add wordAtCursor
This is not actually the way to do it, and can break on some crafted
strings.  The way to actually do it: implement a reverse word search
iterator, then do next() to find a word break, prev() to find a
_valid_ word start, then next() again to find the valid end of said
word.  Maybe 2+, 2-, 1+ actually.

I can probably write a test to see if the cursor spot is ambiguous,
and apply an extra round if so.  Need to mull the rules over before
making any rash moves.