ReverseWordIterator passes conformance test
Ended up needing a clone of the codepoint iterator, so that WB4 can
ignore points in a matter compatible with backward search.

So I created a special SneakIterator which can return WBPs directly,
so as to skip ignorables.  This is also needed for flag emoji, since
the odd-number case must be handle immediately.  So we count back in
a WB4 compatible way, then store the count on the word iterator, and
voila.