Add iterateBefore and iterateAfter
These create reverse or forward iterators before or after a Word.  So
this way, the user can get the word at an index, then iterate forward
or back from that word.

Also:

Fixes #59

Which was fixed awhile back, but I don't feel like doing repo surgery
to tag the fix where it happened.  We have blame for that kind of
thing.